CHANNELMGR-1391 allows correct date manipulation through picker and the two time fields + field styling.
The md-date-picker expects a javascript Date object and returns one when a new date is chosen. The complete correct date values are maintained in a moment object. This includes the time part.

Because the set-to-now button is not bound to an element with an ngModel, we set the form to dirty to enable the save button.

The four display elements of a date field should fit on one line. The styling makes sure that that is ok when the right side panel is at its smallest size. When the panel is bigger, the four elements align together to the left and do not stretch the whole width of the panel.

When the set-to-now button gets focus the icon gets a round grey background circle. The icon is positioned in the middle of this circle.