CHANNELMGR-1391 enhances DateValue class to deal with empty date fields
When creating a new document the date fields are initially empty. When a value is entered in one of the fields of the dateField, the DateValue is populated.

When a document is saved with an empty dateField, the property is not set. This is a difference from what the editor in the content perspective does! To be continued...