CHANNELMGR-1391 recognizes the default empty value of a Date property. Sets property to this value when no date is filled in.
A Date property has a default value representing an empty property. If such a property is read from the backend the frontend field is given an empty string so the field appears empty.

When saving a Date field that has no value (which is allowed when it is not required), the value for the property is set to the default string representing an empty date. This makes that the property is actually created on the document node.