HSTTWO-4268 Serialize the entire page structure instead of only containers
Instead of serializing only containers and container items, we now
serialize the entire page: This way, the page model api pipeline can be
useful for other usages than SPA as well. Of course, we can now get a
deeply nested component hierarchy, but for SPA sites, it is most likely
that the customer will below his root component directly create some
containers and that is it.

Note that I now expose as 'type' of the component the component type
(container, item, or component) and as 'class' the FQCN