HSTTWO-4274 Add 'site' link for content items
This is part 1 of HSTTWO-4274. Admittedly, I do a bit of dirty pipeline
name check, certainly because this is in the content-beans module.

However, for now, I do not see any easy alternative. The more formal
approach with getMount(String alias) will be very complex in usage when
dealing with multiple sites. Hence, I chose for this, arguably, hacky
short cut which in practice most likely will be much more effective than
the eloquent getMount(String alias) approach (since that usage is just
very complex when for example requiring mapped aliases).

Certainly if we will auto-append the PageModelPipeline mount based on some
property, this approach is way more stable