HSTTWO-4274 Enhance link information in page model api content
A content entry can belong to a part of the channel that is not run by
the current SPA. Of course, this is not relevant if the mount has
configure SpaSitePipeline since then the entire mount belongs to a single
SPA. However, if the SpaSitePipeline is configured on some sitemap items
only, it becomes a bit more complex: Is the link to the content entry
possible to show in the current SPA (and possibly fetched via an XHR call),
or can the link only be served by a document request (full html web page).

The page model api now provides this information by the 'type'. If the
type = 'internal', it means that current SPA can serve the content link.

An end project can group the set of sitemap items that are part of the
same SPA together by giving all those sitemap items the same
hst:applicationId.