HSTTWO-4275 Remove the AutoReloadAppender from the generic appenders
The AutoReloadAppender is only needed in the aggregation valve, and for
example not other valves like SelectiveRendering valve. Hence factor
out the AutoReloadAppender from composerComponentWindowResponseAppenders.

This makes the composerComponentWindowResponseAppenders usable by other
pipelines than the DefaultSitePipeline again, and downstream projects
adding an appender will result in this appender being available in all
pipelines. If a downstream project really only wants to inject an
appender that only runs in the DefaultSitePipeline, such a downstream
project can inject the appender then in
defaultAggregationValveComposerComponentWindowResponseAppenders.
Admittedly a long bean name but I did not know a better one