HSTTWO-4281 Expose links for menu items correctly
As a result, I moved the LinkModel object to other package and module.

Instead of serializing the getHstLink and getExternalLink I chose to use
a virtual property since that gives more flexibility to serialize it the
way I want. I could use the HstLinkConverter as well, but I think this
works better