CHANNELMGR-1391 take the user selected time zone into account to display the date
The moment in time as defined in the Date property must be show as that moment appears in the selected time zone. The hours and minutes part of the Date can be affected, but also the day/month/year part, dependant on the position of the dateline.

Introduced the angular-cookies module for easy reading of the cookie value.