CHANNELMGR-1430 Set focus on image element
If no image has been picked yet, we have to set focus on the img element
during rendering; for this we use the focus-if directive.
Otherwise we can set focus on the existing img element directly.
When cancelling a picker we check if there is an image and give it
focus, otherwise we focus the select button.