CHANNELMGR-1765 fix styling keep the left side panel in the right size and position
The left side panel opens and closes by actions implemented in angular material design (mdSideNav) which is different from how the right side panel is opened and closed. Just leave the animation of the left side panel to material design, but make sure the width is set and stored correctly.