CHANNELMGR-1391 Read time zone from user session instead of a cookie
The DefaultLoginPlugin puts the selected time zone in the server-side
user session as well as in a cookie. The cookie is used to re-select
a previously selected time zone in the login page, but was now also used
to initialize the time zone in the Channel Manager. That's an unexpected
side channel between modules that should be avoided.