HSTTWO-4255 Support content rewriting for the PageModelApi
I introduced a custom HippoHtmlBeanSerializer which takes care of
serializing a HippoHtmlBean. It serializes it now as

{
 type = html
 value = ......
}
where the value is the rewritten content. A anchor <a> tag will also
get a data-type attribute indicating whether the link is internal or
external : With this info the SPA can decide to make it an XHR request
or document request.

I have two TODOs
1) If a link cannot be rewritten, we can remove the <a> tag and only
keep the text. Do we prefer that or do we want a broken link?
2) We can remove the HippohtmlBeanMixin because it is not used any more.
Do we want to keep it or not?