CHANNELMGR-1391 set and keep seconds and milliseconds of a Date to zero
Even if the date value string (from the backend) contains values for seconds or milliseconds (other than zero), these values are set to zero. This is in line with how the editor in the content perspective handles this. When initializing a new date field these values are also set to zero.

Added unit tests for this.