CHANNELMGR-1391 make hint and set-to-now work for empty field
If a date field is empty, the label sinks down 'into' it. A click on the input field of the date didn't work, so it got a higher z-index than the label. But when the label is behind the input field the tooltip with the hint doesn't show up. The fix is to place the label still in front of the input field, but with 'pointer-events: none' it does not process mouse events like clicking. So in fact the input field is clicked and gets focus on click. The icon gets 'pointer-events: all' so the hover effect that makes the tooltip appear works. There is one small downside: a click on the icon does not focus the field.

Also fixed a lint error in hippo-cm.config.js.