CHANNELMGR-1391 fix failing unit test
The time zone on the build server is clearly different that on my developer machine. It created the correct date (or moment in time), but the output string was formatted differently due to the time zone ('2018-03-12T07:01:00.000+00:00' instead of '2018-03-12T08:01:00.000+01:00'). Better then to check the specific (milli)seconds parts on the moment object.

Also fixed a linter warning.