CHANNELMGR-1802 Recognize SPA and let it trigger page analysis
A single page application (SPA) can define a window.SPA object to
signal its presence to the Channel Manager. When found, the Channel
Manager does not analyze the page directly when the iframe has loaded,
but first calls SPA.init(). The public JavaScript API of the
Channel Manager is passed as an argument to the init() call. The SPA
can then render/modify the page and trigger (or re-trigger) page
analysis by calling the analyzePage() function in the public API.

Or in a pseudo-interaction format:

1. SPA defines window.SPA
2. iframe onload event triggers
3. Channel Manager detects window.SPA and calls window.SPA.init(cms)
4. SPA renders
5. SPA calls cms.analyzePage()