REPO-1980 Support to reuse the user jcr session across multiple requests
Store the jcr user session as an http session attribute and reuse it
across requests. Only when the http session is pasivated or logged out,
we logout the user jcr session. This avoid potentially frequent logins
for the same jcr user session. Certainly when there are many domain rules
logging in can be more CPU expensive. Or when there is for example
external LDAP authentication involved, the i/o of logging in a session
might become too large. This fix improves those scenarios