CHANNELMGR-1390 also take user timezone into account for date-only field
Just setting the hours and minutes to zero is not enough, timezone must be taken into account. Using the moment function startOf('day') is the best way, and must be done after setting the time zone.

Also some fixes to make the markup simpler and using a default value for the dateOnly parameter of the dateValue model object.