HSTTWO-4257 Only add the page model api mount if it should be added
When for example adding on hst:virtualhost level that the hst:pagemodelapi
is available below 'resourceapi', then mounts that are not mapped of
course still cannot be a resource api: Hence we should skip those.

Also if a mount does not have channel info (aka is typically a rest mount
for example a content rest mount), then it does not make sense to add a
page model api mount below it.