CHANNELMGR-1803 Render component in SPA before doing it the old way
When the SPA defines a 'renderComponent()' function that function is
called when a component needs to be (re)rendered. When it is missing or
returns false, the Channel Manager render a component as it did before.

The component rendering logic moved to a new ComponentRenderingService,
which orchestrates to call to the SPA or fallback to the
PageStructureService. The ComponentRenderingService now also handles the
render-component event from ExtJs, so all component rendering logic is
located in one place.

The RenderingService has been renamed to ChannelRenderingService since
there's now also a ComponentRenderingService.