REPO-1980 Support more generically to store a jcr session on an http session
I plan to move the HttpSessionBoundJcrSessionHolder to a generic
location next such that also other webapps can reuse this facility object.

The valueUnbound has been made easier after review comments of Peter.

Also I extracted a getOrCreateJcrSession which does all the logic of
correctly binding the newly created jcr session on the http session or
return a previously bound jcr session. It takes the userId into account
because when invoking getOrCreateJcrSession for the same http session
but with different credentials, a different session should be returned
of course.