CHANNELMGR-1803 Move component rendering logic to HippoIframeCtrl
The ComponentRenderingService was not created because no other piece of
the app depended on it. To get it running we would need a separate
'run' function, which is ugly. Having the controller handle the external
event is much more logical and nicely ties it to the lifecycle of the
hippoIframe.