HSTTWO-3883 Make sure that any jcr sessions bound to earlier http session get logged out
In the comment I added the reason: When a cms user logs out and then in
again, the HST http session survives. This means, it can still have jcr
sessions bound to its http session which are stale (the cms user logged
in and out possibly updating its authorization rules). Hence, in such cases,
we need to clear all bound jcr sessions for a certain attribute namespace

Also removed an outdated html comment which was describing that every
request did get a fresh new jcr session