CMS-11158 Support for un-bounding jcr sessions by attribute prefix
Typically, when a cms user logs out and in again, the HST http session
survives. Any bound jcr sessions in the HST http session then typically
need to be unbound, so they are recreated (for example important when a
cms user logs out and in to update his/her changed authorization rules:
then of course, we also want to log out the HST channel mngr user for this
cms user). This method facilitates that