HSTTWO-4297 Cross Mount links should get data-type external
when the Mount of the resolved Link is different than the parent mount
of the current request (PageModelApi mount), it means the link is external.

It can very well be that the backing sitemap items of the current matched
sitemap item and the resolved link are both null, but the mounts are
different. Hence this check should of course first be done