HSTTWO-4297 Support a target mount for links including cross mount fallback
for the content rewriters there is never a reason to not want the fallback
when an explicit target mount is provided...it would always result in
that case in a not found link if the target mount could find a link.

Therefor I changed this behavior and this won't impact any existing
rewriters negatively. I could also get rid of java comments warning about
the target mount to be null....which was a red flag already. Now it makes
more sense