ESSENTIALS-1188 exclude most transitive dependencies from Maven archetype-common dependency
The archetype-common dependency was added with ESSENTIALS-1112 to let Maven provide the readPom and writePom implementation.
However archetype-common pulls in a huge number of additional Maven dependencies (pretty much 'everything Maven') which isn't
needed/used for just the reading/writing of pom files.
Those unnecessary transitive dependencies are now excluded, except for maven-core and maven-artifact which are needed/used for
other features.