ESSENTIALS-1145 add more documents and links between detail and overview pages
There is one author with 3 blog posts now, which makes the numbers in the facets vary a bit more, to give the user more insight in how facets work. Also on the detail page of a blog of this user you can now see data in the "more by..." section on the right.

More category entries have been added and selected in the different documents. This also make the numbers change in the facets.

To make it more clear that there is a detail page to go to, there is now a link on the overview pages. This is not done on the JSP templates.