CHANNELMGR-1814 Color icons for the publication state of a document
The Material icons have thicker lines, so the color of the black and
white outlined icons looked very harsh compared to the custom SVG icons
next to it. Carolien suggested to use colored icons instead, which
looked quite good. We use the 'filled' variant of the icon, so they look
the same as the ones used in the content perspective.