CHANNELMGR-1814 Use the MDI alert icon to indicate channel changes
Removed the need for our custom SVG icon. One snag: the exclamation mark
in the icon is transparent, so on the gray toolbar it becomes gray. That
looked quite ugly. Material icons don't allow multiple colors, but with
some CSS trickery the exclamation mark has been forced white again so it
looks consistent with the icon in the content editor.