CHANNELMGR-1819 Add confirmPublication and publish methods
Show a 'publish' confirmation dialog for pristine documents.
Show a 'save-and-publish' confirmation dialog for dirty documents.

The publish function first deletes the draft, then executes the workflow
action, and finally creates a new draft. If either createDraft or
deleteDraft fails, an error is displayed and the promise is rejected.
When the workflow action fails, we still try to create a new draft to
let the user continue editing, and afterwards reject the publish
promise.
When the workflow action succeeds, and info message is displayed.