CHANNELMGR-1815 retrieve publish workflow hints from a document that is not in edit mode
The workflow is designed so that when in edit mode, no publish actions are allowed. But for the editor in the channel manager we must now this state while also offering edit options. The workflow hints must be retrieved from the workflow of a document that is not in edit mode (obtainEditableInstance has not been called yet).