CHANNELMGR-1840 Provide hardcoded 'test' extension to the channel editor
In the future, CmsExtensionLoader should have different implementations.
For now we hard-code a single 'test' extension.

The WicketExtJs framework uses the org.json library for passing JSON
properties to the frontend. However, we'd like to serialize the object
model for CmsExtensions to JSON, which is easily done via Jackson.
With the added jackson-datatype-json-org a Jackson object mapper can
convert a Java object model to org.json objects.

The 'test' extension is supposed to load a channel, hence the URL path
'/site/_cmsinternal/testextension'.