CHANNELMGR-1821 Use 'bump' animation instead
Animating the color turned out to be quite tricky since the color of
the button can differ (normal or hovered). The new 'bump' animation
makes the button slightly bigger for a short moment.

Animating the button itself was tricky because the animation on the
button also ran when the button was removed from the DOM, which delayed
the actual removal. As a result, both the Publish and the Save button
were visible. The solution was to only define the animation when the
button got added to the DOM by using the 'ng-enter' class set by
ngAnimate.