CHANNELMGR-1843 Show tab per page extension in right side-panel
The PageInfoService registers a separate ui-router state for each page
extension. Clicking a tab activates that state, so a state is only
loaded when the tab is clicked. Each of these states is also 'sticky',
so the sticky-state plugin of ui-router remembers the client-side state
of each tab as long as one of the tabs is active. The caveat here is
that each of the sticky states need to target a different ui-view, since
that's the place where the sticky-state plugin remembers the client-side
state. Hence the two ng-repeats in the pageInfoMain template: one for
the tabs, and one for the ui-views for these tabs.

For now the 'path info' of the current page is passed to the page
extension tabs. This should be changed to the proper full page URL(s).