CHANNELMGR-1827 Improve detection of RequireJs
Checking for the existence of a window.require function is not enough
since some customers have defined such a function without using
RequireJs. The drag-drop service now also checks whether the
window.require function is the same as the window.requirejs function
(RequireJs defines both). That should be enough to be certain about
RequireJs usage.

Using iframe.frameElement.contentWindow was a bit convoluted: 'iframe'
is already the window object of the iframe, so just use that. There's
also no need to keep a reference to the require function.