CHANNELMGR-1363 Fix broken scrollbar tests on osx
On macOs, the default setting for scrollbars is to only show them when
scrolling. When using this setting, the scrollbar overlays the page and
does not consume any space. The tests should be aware of this and adjust
the test values accordingly.
This only applies to the vertical scroll values because the horizontal
scrollbar is rendered outside of the iframe, thus reducing the vertical
viewport which in turn makes the amount of pixels to scroll higher.

To make the test more readible I added variables and comments.