CHANNELMGR-1898 Don't use preview channel version for test extensions
Rendering a channel via the _cmsinternal prefix changes backend state.
From that moment on, the HST will assume the rendered channel is the
'current' one. This is not what we want for the test extensions:
rendering a test extension should keep the 'current' channel.

The solution is to use the 'live' URL of the test extension channels
instead, which on localhost:8080 is /site/test-extension.