CHANNELMGR-1896 Do not delete 'unsupported' choice field nodes
When writing the supported values of a choice field, we merge in the
unsupported field values with a marker instance UNSUPPORTED_FIELD_VALUE.
When calling #writeValue with the UNSUPPORTED_FIELD_VALUE instance, we
simply do nothing.
This way we no longer need to remove the unsupported value nodes and
still have correct behavior like cardinality checks.