CHANNELMGR-1913 Configure extensions in the repository
Extensions are configured as individual nodes below
/hippo:configuration/hippo:frontend/extensions, which seemed the most
logical location that did not require a CND change. The 'extensions'
node can be of type nt:unstructured.

The name of an extension node is its ID. Each extension node should
define three string properties:
- displayName for the display name
- context for the string-version of the context (which can only be
  "page" for now since we won't support other extensions yet)
- urlPath for the URL path load the extension from

The CmsExtensionValidator performs some sanity checks on the values of
these properties, so developers get quick feedback when they add odd
values or forget to add properties. Separating the loader from the
validator allows future reuse of the validator for different loaders.

The JcrCmsExtensionLoaderTest uses MockNodes, which required a
dependency on repository-mockutils.