CHANNELMGR-1913 ids of extensions must be unique. Config path one level deeper.
Added code and tests to make sure no extensions with duplicate IDs exist. Added a test for it. Placed the config of the extensions one level deeper inside the repository, inside the /cms scope.

Updated java doc and a log message. Also made it a warning instead of an info message.