CHANNELMGR-1974 Add mask while resizing to prevent loss of mouse events
When resizing a sidepanel, we used a couple of CSS workarounds to
prevent unwanted side-effects:
- user-select: none -> prevent text-selection
- pointer-events: none -> prevent loss of mouse events when hovering an
 iframe

While this works, it is not trivial and needs code changes if we want
to introduce iframes in side-panels.

The simpler way to fix it is; render a transparent mask covering
the whole app while resizing, the mask captures (and blocks) all mouse
events. By listening for mousemove/mouseup events on the mask instead
of the document we make the whole process a bit more lean and
encapsulated.

After some dialog with Carolien, we decided to drop the showing/hiding
of overlay elements as well, as it is more in line with all other
resizing options (open/close sidepanel, device switcher, window resize).

The only real issue with this approach is the fact that we need to set
the attribute draggable=true on the handle-inner element; this prevents
Firefox from still bleeding mouse-events through the mask.

(cherry picked from commit c7a6403)