CHANNELMGR-1853 Add mask while resizing to prevent loss of mouse events
When resizing a sidepanel, we use a couple of CSS workarounds to prevent
unwanted side-effects:
- user-select: none -> prevent text-selection
- pointer-events: none -> prevent loss of mouse events when hovering an
 iframe
This needs to be extended for OpenUI as we are adding even more iframes,
but there is a simpler way to fix it; render a transparent mask covering
the whole app while resizing, the mask captures (and blocks) all mouse
events. By listening for mousemove/mouseup events on the mask instead
of the document we make the whole process a bit more lean and
encapsulated.

After some dialog with Carolien, we decided to drop the showing/hiding
of overlay elements as well, as it is more in line with all other
resizing options (open/close sidepanel, device switcher, window resize).

The unit tests were lacking quite a lot, so those had to be implemented
as well.

The only real issue with this approach is the fact that we need to set
the attribute draggable=true on the handle-inner element; this prevents
Firefox from still bleeding mouse-events through the mask.