REPO-2019 Simplify GetBranchTask and use lowercase for state
The state suffix for the labels was uppercase, but the Variant enum in
the WorkflowUtil also contains a state String field, which is identical
to Variant#name().toLowerCase(). So to reduce confusion in naming I
made the state method in the enum public and used it for labeling of
the versions.

In addition I simplified the code of the GetBranchTask.