CHANNELMGR-1906 add tooltip to tab. toggle tab pagination after resize.
If tab names are long and tabs don't fit the width of the panel, names get ellipses. To have a title attribute (for a browser tooltip) is then a good thing. After a resize of the panel the pagination handles for the tabs may need to be shown or hidden. This is triggered by dispatching a window resize event.