CHANNELMGR-1916 Separate page-specific logic from iframeExtension
The iframeExtension is now a more abstract and reusable component. It
updates an extension in an iframe whenever the bound context changes
using the $onChanges lifecycle hook. It can be reused in the future for
additional iframe extensions for different contexts.

The initial context is set once the iframe has been loaded. Meanwhile,
the context could already change. To avoid updating the context too
soon, the iframeExtension now keeps track of whether the iframe has
already been loaded.

The pageExtension component is loaded for page-info states. It uses the
ui-router state parameters to load the right iframe extension.