CHANNELMGR-1916 Append extension URL path to CMS location
Extensions will - for now - be packaged in the cms.war file, and are
therefore served via the same origin as the CMS. The URL per extension
can therefore be relative (i.e. without an origin). But the CMS may use
a context path (e.g. '/cms' on localhost and usually '/' in production).
The URL path of an extension should therefore be prefixed with the
CMS context path.

Since the extension URL does not change once the iframeExtension
component has been loaded, the ng-src value can use one-time binding to
prevent a watch.