CHANNELMGR-1904 Don't show publish button for document added to project
Reload content editor once document is added to project so the document
is re-fetched from the backend. The backend then updates the canPublish
and canRequestPublication flags based on that are set by the backend
based on whether the 'publish' or 'requestPublication' workflow
actions are available (which they aren't once the document is part of
a project). Without the reload, the initial load of the document will
indicate that these workflow actions are available, and that information
won't be updated after the document had been added to a project.