CMS-10665 Set link-target to null if selected node is not valid
Previously, if a valid node was selected and later an invalid one was
selected, the link-target would still point to the valid node. This
caused issues with, for example, the RTE link picker when toggling the
"open in new window" checkbox; it would re-enable the "OK" button while
the selected node was not valid, because the link-target still
referenced a valid node.

Had to add a couple of null checks to fix NPE because of this fix.