CHANNELMGR-1931 Skip problematic property copy-computed expectation
Chrome v67 exhibits weird behavior around the '-webkit-app-region' CSS
property: when it is copied it changes the value from 'none' to
'no-drag'. This property is only used in Chrome apps (like Electron
based apps) and can safely be skipped in the validating unit test.

The related unit test now check every single property instead of the
cssText property to give a more detailed error when the test fails.