ESSENTIALS-746 Fix problem with copy-action of directory instruction
The copy action of the directory instruction had 2 problems:

1) if the source directory didn't have a unique name among all JARs
packaged in the essentials.war, any matching JAR was used to copy from

2) the matching between the source directory and candidate entries in
the JAR failed due to different path formats.

Issue 1) is solved by selecting the appropriate JAR based on the plugin's
packageFile (which is expected to be unique per JAR, and which triggers
the execution of the directory instruction). To access the packageFile,
the plugin's descriptor is now passed into the instruction execution as
one of the execution parameters. In a next major version, this API could
be made less "magic".
Issue 2) is solved by normalizing the directory path, relative to the
JAR's root.