CHANNELMGR-1894 Call _onActivate when selecting ChannelEditor card
When a user click on "view ChannelX" from the document perspective, we
rely on the following order of client side calls to actually show the
related document:

- initChannel()
- loadChannel()

If the ChannelEditor has not been rendered yet (channel-overview is
active), rootPanel._onActivate() will never call loadChannel().
This means we have to ensure the ChannelEditor card is selected prior
to rootPanel._onActivate() being called, but that never happens; the
"CMSChannels" event is executed before the ChannelEditor card is
selected and this we see a blank screen.
To fix this, we now call rootPanel._onActivate when the ChannelEditor
card is selected.

I also changed the order of calls in
ChannelManagerPerspective#viewChannel. It does not have any impact on
the order of execution in the client side, but reads more logically.