CHANNELMGR-1894 Fix first view of page in unactivated Channel Editor
The _onActivate is called before the card is selected, because the
JavaScript for the latter is only rendered when the RootPanel is
rendered, which happens after the _onActivate event has been generated.
Fixed by adding RootPanel#activateCard(), which activates a given card
'directly' by rendering some JavaScript for it. That was the JavaScript
that selects the Channel Editor card is run before the JavaScript that
generated the _onActivate event.