CHANNELMGR-1894 Fix second view of document in channel with mount path
The 'initialPath' provided by view-document-in-channel is relative to
the channel's mount. However, the load-a-path-in-the-current-channel
logic in the hippoCm controller did not prefix the initialPath with the
mount path of the current channel. Hence the loaded path was matched in
the channel for the root mount, which caused a channel switch.

Fixed by using the same logic to initialize the initialRenderPath in the
ChannelService (i.e. prefix the mount path) for loading a path in the
current channel.