HSTTWO-4365 communicate which site should be rendered between webapps
Since the platform webapp owns the page composer (the jar that supports
the modifications to the HST model), all channel manager related calls
that are not about really rendering the preview webpage go via the cms
webapp. Therefore we cannot just use the http session any more in the page
composer jar to set information about what is the preview site that the
channel mgr should render.

As a workaround I now use the payload from the CmsSessionContext to share
info about which preview to render in the channel mgr. I added TODOs to
revisit this choice later.

Also we need to revisit the 'getCmsPreviewPrefix' whether we can still
support it as we did.

Lastly, still need to validate it all works behind a proxy (for example
httpd where the client does not 'see' the context path)