CHANNELMGR-1954 Add router state for add to project confirmation
The user must be presented with the option to add a document to a
project if it is not associated yet before an editable instance is
created in the back-end. If the user cancels, then the rightside panel
should close and no project version of the document is created.

This is solved by first getting the document from the back-end.
If the branchId of the document is not the same as the id of the
selected project then the add-to-project ui-router state is invoked
so that the user can proceed or cancel.