CHANNELMGR-1946 Store context path of current channel in HstService
Storing it in the ConfigService was an ugly workaround to avoid a
circular dependency between services. Storing it in the HstService makes
much more sense since it's the only place where it's used.

To avoid too many method parameters we now pass the whole channel to the
initialize() methods of SessionService and HstService.