CMS-11289 refactored and simplified HippoServiceRegistry as a singleton service registry only
The @SingletonService and @WhiteboardService annontation now no longer supported and thus dropped.
Instead, two new WhiteboardServiceRegistry and WhiteboardProxiedServiceRegistry, together with
corresponding ServiceTracker and ProxiedServiceTracker interfaces, to re-implement this capability
in a cleaner and more comprehensible way. See the javadoc for each and all of these classes.

Note: this is a *major* breaking change requiring to adjust almost all HippoServiceRegistry usages,
except for 'plain' singleton service registrations and lookups.