HSTTWO-4376 Make sure that on the first request the preview is shown
when a channel is opened from the Channel Manager overview,
then when opening the channel, in the URL 'render host' information, is
send. Thus the URL will be something like:

http://cms.example.com:8080/site/_cmsinternal/
       ?org.hippoecm.hst.container.render_host=www.example.org

This request will use 'www.example.org' instead of 'cms.example.com'
to match the request.

The HST however in the HstDelegateeFilterBean the request will then be
marked to come from the CMS (channel mng request).

Next, in the CmsSecurity valve, it will be detected that the visitor
does not yet have an SSO between cms and site webapp.
It will create one via the CmsSessionContext. Some redirects for the
handshake will take place, after which the preview of the
channel will be loaded in the channel mngr